ugly hack to crash the browser despite Elm 0.19
@sgraf812 proposed to use an infinite loop instead of `Debug.crash`
because in these cases there are no sensible defaults and it is
better to make the browser hang (or hopefully error out) than to return
incorrect data.  Also, hopfully this path will never be executed.

Nonetheless, I feel uncomfortable about putting this in, but I'm not
sure what the alternative is.